JS Done
JS code creates the heading for the the page's state's elections. JS creates tables with each candidate, their party, the number of votes they got, and the total percentage of votes they got. A function creates a table to contain the information for each. The function also divides the candidate votes by the total votes, returning what it is to the tenth decimal. A function also creates a bar next to the information, which is based on the percentage that is gotten.